Rename Heap to MaxHeap, move to package Heap and update relevant usages
